Separate HLS audio and video streams
 Allows:
  * The HLS player to propose an "Audio only" resolution
  * The live to output an "Audio only" resolution
  * The live to ingest and output an "Audio only" stream

 This feature is under a config for VOD videos and is enabled by default for lives

 In the future we can imagine:
  * To propose multiple audio streams for a specific video
  * To ingest an audio only VOD and just output an audio only "video"
    (the player would play the audio file and PeerTube would not
    generate additional resolutions)

This commit introduce a new way to download videos:
 * Add "/download/videos/generate/:videoId" endpoint where PeerTube can
   mux an audio only and a video only file to a mp4 container
 * The download client modal introduces a new default panel where the
   user can choose resolutions it wants to download

+++ scripts/build/server.sh
@@ -2,7 +2,9 @@
 
 set -eu
 
-rm -rf ./dist ./packages/*/dist
+if [ -z ${1+x} ] || [ "$1" != "--incremental" ]; then
+  rm -rf ./dist ./packages/*/dist
+fi
 
 npm run tsc --  -b --verbose server/tsconfig.json
 npm run resolve-tspaths:server
